+++ analysis/DataLoader.hh
@@ -48,15 +48,18 @@ public:
 	     bool        debugIn           = false,
 	     bool        processSamplersIn = true,
 	     bool        allBranchesOn     = true,
-	     const RBDS::BranchMap* branchesToTurnOn = nullptr);
+	     const RBDS::BranchMap* branchesToTurnOn = nullptr,
+	     bool        backwardsCommpatible = true);
   virtual ~DataLoader();
 
   /// Create an instance of each class in the file to be overlaid by loading
   /// the ROOT file.
-  void CommonCtor(std::string fileName);
+  void CommonCtor(std::string fileName,
+		  bool backwardsCommpatible);
 
   /// Build up the input file list.
-  void BuildInputFileList(std::string inputPath);
+  void BuildInputFileList(std::string inputPath,
+			  bool backwardsCommpatible);
 
   /// Open the first file in the file list and map the trees in it.
   void BuildTreeNameList();